#1: Menu added

+++ lionengine-core/src/com/b3dgs/lionengine/drawable/Sprite.java
@@ -99,7 +99,7 @@ public interface Sprite
      * 
      * @param alpha The alpha value <code>[0 - 255]</code>.
      */
-    void setAlpha(byte alpha);
+    void setAlpha(int alpha);
 
     /**
      * Get the current sprite width (its current size, after scaling operation).
